% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDataType.R
\name{ChangeDataType}
\alias{ChangeDataType}
\title{Change Data Type}
\usage{
ChangeDataType(InputValueType, TargetValueType, data)
}
\arguments{
\item{InputValueType}{data value type of input data}

\item{TargetValueType}{data value type the output data should be}

\item{data}{the input data set}
}
\value{
\code{data} the output data set with wanted data value type
}
\description{
Change data value type between "Beta" and "M". As we have:
B = 2^M/(2^M + 1) and M = log2(B/(1-B))
}
