% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetsampleSet.R
\name{GetsampleSet}
\alias{GetsampleSet}
\alias{Getsentrix_barcode}
\alias{Getsentrix_position}
\title{Get sample Set from rgSet}
\usage{
GetsampleSet(rgSet)

Getsentrix_barcode(rgSet)

Getsentrix_position(rgSet)
}
\arguments{
\item{rgSet}{The RGChannelSetExtended with in the list \code{loading}}
}
\value{
\code{sampleSet} sample set of the raw data

\code{sentrix_barcode} the sentrix barcode of samples

\code{sentrix_position} the sentrix position of samples
}
\description{
\code{GetsampleSet} is used to extract the samples set from
the rgSet. \code{Getsentrix_barcode} is used to extract sentrix barcode from
rgSet. \code{Getsentrix_position} is used to extract sentrix position from rgSet
}
