% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeData.R
\name{imputeData}
\alias{imputeData}
\title{Imputation of missing value}
\usage{
imputeData(inputdata.small)
}
\arguments{
\item{inputdata.small}{input data set which needs imputation of
missing values}
}
\value{
\code{inputedata2} the data set after imputation, without missing
values now
}
\description{
K-nearest neighbor averaging (KNN) is used to impute missing vlaues
of the input data.
}
