% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimData.R
\name{TrimData}
\alias{TrimData}
\title{Trim data set}
\usage{
TrimData(probeNames, inputdata)
}
\arguments{
\item{probeNames}{the probe list needed for prediction models}

\item{inputdata}{input data}
}
\value{
\code{inputdata.small} The smaller data set with only specific
probes included
}
\description{
Trim the data set with probes for specific prediction models
}
