% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.BMIQ.R
\name{UniD.BMIQ}
\alias{UniD.BMIQ}
\alias{Apply.BMIQ}
\title{Type II probe normalization-BMIQ}
\usage{
UniD.BMIQ(Beta.raw = Beta.raw, arrayType, outDir, write)

Apply.BMIQ(beta.v, design.v, nL = 3, doH = TRUE, nfit = 50000,
  th1.v = c(0.2, 0.75), th2.v = NULL, niter = 5, tol = 0.001,
  plots = FALSE, sampleID = 1)
}
\arguments{
\item{Beta.raw}{The raw beta value data frame generated by
\code{UniD.dataqc}, with missing value.}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{outDir}{directory where output data should be saved if write = T}

\item{write}{whether the output should be saved, highly recommended}

\item{beta.v}{vector consisting of beta-values for a given sample}

\item{design.v}{vector specifying the probe Type (I = 1 or II = 2)}

\item{nL}{number of states in beta mixture model, default is 3}

\item{doH}{perform normalisation for hemimethylated Type II probes,
default is T}

\item{nfit}{number of probes of a given design to use for fitting,
default is 50000}

\item{th1.v}{thresholds used for the initialisation of the
EM-algorithm, default is between 0.2 and 0.75}

\item{th2.v}{thresholds used for the initialisation of the
EM-algorithm, default is NULL}

\item{niter}{maximum number of EM iterations to do, default is 5}

\item{tol}{tolerance threshold for EM algorithm, default is 0.001}

\item{plots}{whether plot the figure, default is FALSE}

\item{sampleID}{the ID of the sample being normalised, default is 1}
}
\value{
A data frame with the normalized data
}
\description{
Used to normalize the data generated from 450k
or higher platform. Due to the two types of probes used in those platforms,
the beta value generated from the Type I and Type II probe is not directly
comparable. For the gene expression subtype prediction, data needs to be
normalized using BMIQ.
\code{Apply.BMIQ} was utilised within \code{UniD.BMIQ} with all default
settings. No need to change.
}
\section{Functions}{
\itemize{
\item \code{Apply.BMIQ}: Apply the BMIQ normalization
}}

\examples{
\dontrun{
Beta.BMIQ <- UniD.BMIQ(Beta.raw, outDir = "~/Desktop/output/",
arrayType = "EPIC", write = T)

Beta.BMIQ <- UniD.BMIQ(Beta.raw,  outDir = NULL,
arrayType = "EPIC", write = F)
}
}
\references{
Teschendorff, A. E., et al. (2013). "A beta-mixture quantile normalization method
 for correcting probe #' design bias in Illumina Infinium 450 k DNA methylation
 data." Bioinformatics 29(2): 189-196.
}
\seealso{
The BMIQ normalization method is refer to \code{\link[wateRmelon]{BMIQ}}
}
