% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.dataqc.R
\name{UniD.dataqc}
\alias{UniD.dataqc}
\title{Probe quality evaluation}
\usage{
UniD.dataqc(loading = loading, outDir, detP.cut = 0.05, bc.cut = 3,
  arrayType, write)
}
\arguments{
\item{loading}{List generated from the \code{UniD.load()}}

\item{outDir}{directory where output data should be saved if write = T}

\item{detP.cut}{The cut point for detection P-value, default is 0.05.
If the probe's detection p-value > 0.05, the probe will be set as missing
value in a sample-wise fashion.}

\item{bc.cut}{The cut point for beadcount per probe, default is 3.
If beadcount < 3, the probe will be set as missing value sample-wise.
control probes}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{write}{whether the output should be saved, highly recommended}
}
\value{
A data frame with beta values for all samples and probes.
}
\description{
Used to evaluate the probe quality in term of
probe detection p-value and beadcount. For probes which failed the
cut point, they will be set as missing value. The data with missing
value will be returned/saved as beta value.
}
\examples{
\dontrun{
Beta.raw <- UniD.dataqc(loading = loading, outDir = "~/Desktop/output/",
detP.cut = 0.05, bc.cut = 3, arrayType = "450k", write = T)

Beta.raw <- UniD.dataqc(loading = loading,  outDir = NULL, detP.cut = 0.05,
bc.cut = 3, arrayType = "450k", write = F)
}
}
