% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.intctl.R
\name{UniD.intctl}
\alias{UniD.intctl}
\title{Checking quality of internal control probes}
\usage{
UniD.intctl(loading = loading, dataDir, outDir, arrayType, sampleType, write)
}
\arguments{
\item{loading}{List generated from the \code{UniD.load()}}

\item{dataDir}{directory where raw data stored}

\item{outDir}{directory where output data should be saved if write = T}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{sampleType}{Indicate the input data are "FFPE" or "other"}

\item{write}{whether the output should be saved, highly recommended}
}
\value{
a data frame with quality checking results for internal
control probes
}
\description{
Used to check the quality of internal control
probes and return a data frame which includes whether
samples have failed on different categories of internal controls
}
\examples{
\dontrun{
samQC <- UniD.intctl(loading = loading, dataDir = "~/Desktop/IDAT/",
outDir = "~/Desktop/output/", arrayType = "450k", sampleType = "FFPE",
write = T)

samQC <- UniD.intctl(loading = loading, dataDir = "~/Desktop/IDAT/",
outDir = NULL, arrayType = "EPIC", sampleType = "other", write = F)
}
}
