% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.load.R
\name{UniD.load}
\alias{UniD.load}
\title{Load raw data}
\usage{
UniD.load(dataDir, outDir, arrayType, write)
}
\arguments{
\item{dataDir}{directory where raw data stored}

\item{outDir}{directory where output data should be saved if write = T}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{write}{whether the output should be saved, highly recommended}
}
\value{
A list which includes rgSet, Mset, and detP
}
\description{
Used to load the raw data and return a list.
}
\examples{
\dontrun{
loading <- UniD.load(dataDir = "~/Desktop/IDAT/", outDir = "~/Desktop/output/",
arrayType = "450k", write = T)

loading <- UniD.load(dataDir = "~/Desktop/IDAT/", outDir = NULL,
arrayType = "EPIC", write = F)
}
}
