% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.probefilter.R
\name{UniD.probefilter}
\alias{UniD.probefilter}
\title{Probe and sample filter}
\usage{
UniD.probefilter(Beta.raw, outDir, filterXY = TRUE, filterSNPHit = TRUE,
  filterMultiHit = TRUE, filterNonCG = FALSE, filterNonEpic = TRUE,
  arrayType = c("450k", "EPIC"), filterSample = TRUE,
  filterSample.cut = 0.1, filterProbe = FALSE, filterProbe.cut = 0.05,
  write)
}
\arguments{
\item{Beta.raw}{data frame generated from the \code{UniD.dataqc()}}

\item{outDir}{directory where output data should be saved if write = T}

\item{filterXY}{whether filter out probes located on Chromosome
X and Y. Default value is TRUE.}

\item{filterSNPHit}{whether filter out probes may affected by SNP.
Default is TRUE. Probe list adapted from Zhou W, Nucleic Acids Research,
 2017.}

\item{filterMultiHit}{whether filter out probes can mapped to
multi locations. Default is TRUE. Probe list adapted from Nordlund J,
 Genome Biology, 2013.}

\item{filterNonCG}{whether filter out probes which are not targeting
CpG sites. Default is FALSE.}

\item{filterNonEpic}{whether filter outprobes which are available on
450k platform but not available on EPIC platform. Default is T. Highly
recommended if building models with data from 450k platform.}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{filterSample}{whether filter out samples with high proportion
of missing values. Default is TRUE.}

\item{filterSample.cut}{if \code{filterSample} = T, the threshold for
high proportion of missing values per sample. Default is 0.1.}

\item{filterProbe}{whether filter out probes with high proportion
of missing values. Default is TRUE. Carefully usage with small number of
sample size.}

\item{filterProbe.cut}{If \code{filterProbe} = T, the threshold for
high proporition of missing values per probe. Default is 0.05.}

\item{write}{whether the output should be saved, highly recommended}
}
\value{
A data frame with probes after filtering in Beta value
filtering
}
\description{
Used to filter out probes with suspect quality
issues for following categories: (1) probes located on ChrX/Y; (2) probes
may affected by SNP; (3) probes may mapped to multiple locations; (4)
probes not targeted CpG sites; (5) probes on 450k platform but not
available on EPIC platform; (6) probes with large proportion of missing
values.
It can also filter out samples with large proportion of missing
values. Missing values may caused by non-significant detection p-value
or less beadcount per probe.
}
\examples{
\dontrun{
Beta.clean <- UniD.probefilter(Beta.raw, outDir = "~/Desktop/output/",
filterXY = T, filterSNPHit = T, filterMultiHit = T, filterNonCG = F,
filterNonEpic = T, arrayType = "450k", filterSample = T, filterSample.cut
= 0.1, filterProbe = F, write = T)

Beta.clean <- UniD.probefilter(Beta.raw, outDir = NULL,
filterXY = T, filterSNPHit = T, filterMultiHit = T, filterNonCG = F,
filterNonEpic = F, arrayType = "EPIC", filterSample = T, filterSample.cut
= 0.1, filterProbe = F, write = F)
}
}
\references{
Zhou, W., et al. (2017). "Comprehensive characterization, annotation and innovative use of Infinium DNA #' methylation BeadChip probes." Nucleic Acids Res 45(4): e22.

Nordlund, J., et al. (2013). "Genome-wide signatures of differential DNA methylation in pediatric acute #' lymphoblastic leukemia." Genome Biol 14(9): r105.
}
