% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniD.R
\docType{package}
\name{UniD}
\alias{UniD}
\alias{UniD-package}
\title{UniD: A package for unified diagnostic platform for gliomas.}
\usage{
UniD(dataDir, outDir, arrayType, sampleType, write)
}
\arguments{
\item{dataDir}{directory where raw data stored}

\item{outDir}{directory where output data should be saved if write = T}

\item{arrayType}{the platform which raw data generated, can be "450k" or "EPIC"}

\item{sampleType}{Indicate the input data are "FFPE" or "other"}

\item{write}{whether the output should be saved, highly recommended}
}
\value{
A data frame with all predicted results
}
\description{
The UniD package use the DNA methylation microarray data to predict other
genomic information, including IDH mutation, ATRX mutation, TERT promoter
mutation, chromosome 1p/19q co-deletion, TCGA gene expression subtypes (
Classical, Mesenchymal, and Proneural) with their probability, and MGMT
gene expression level.
}
\details{
UniD package providing two categories of functions:

(1) Data processing and quality control:

\code{UniD.intctl}, \code{UniD.dataqc}, \code{UniD.probefilter}

(2) Genomic information prediction:

\code{UniD.pred}
}
\section{Functions}{

\code{UniD.load}: load the raw data

\code{UniD.intctl}: check the internal control probes which can reflect
different aspects of experiments quality.

\code{UniD.dataqc}: check data quality in three aspects: detection P-value,
 beadcount per probes, and missing values.

\code{UniD.probefilter}: filter out probes belong to different categories
which may have quality issues, and probes with high proportion of missing
values.

\code{UniD.BMIQ}: normalize type I and type II probes based on BMIQ method.

\code{UniD.pred}: predict othergenomic information using Illumina DNA methylation
microarray data, currently support 450k and EPIC platform
}

\examples{
\dontrun{
 pred.result <- UniD(dataDir = "~/Desktop/input/", outDir = "~/Desktop/output/",
 arrayType = "450k", sampleType = "other", write = T)
}
}
\seealso{
The BMIQ normalization method is adapted from \code{\link[wateRmelon]{BMIQ}}.
}
